requirements.txt
PyYAML
regex
elasticsearch
pygrok
pandas